refactor: 調整檔案架構，讓整體架構較好理解
將組合較多的元件搬離ui資料夾
並調整檔案的對應路徑
